refactor(polar_alignment): move solver out of __init__.py
PiFinder/polar_alignment/__init__.py held the whole solver; it now
lives in solver.py with a thin __init__ re-exporting the API, so
existing imports keep working.
